Rename stat_data_allocator.h to allocator.h
