+++ source/common/stats/metric_impl.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <vector>
 
-#include "envoy/stats/stat_data_allocator.h"
+#include "envoy/stats/allocator.h"
 #include "envoy/stats/stats.h"
 #include "envoy/stats/tag.h"
