+++ source/common/stats/thread_local_store.cc
@@ -6,9 +6,9 @@
 #include <memory>
 #include <string>
 
+#include "envoy/stats/allocator.h"
 #include "envoy/stats/histogram.h"
 #include "envoy/stats/sink.h"
-#include "envoy/stats/stat_data_allocator.h"
 #include "envoy/stats/stats.h"
 
 #include "common/common/lock_guard.h"
